fix: extract parse sorting function
