+++ src/helpers/sorting.ts
@@ -4,33 +4,12 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { updateSettings } from '@zextras/carbonio-shell-ui';
-import { isTrash, JSNS } from '@zextras/carbonio-ui-commons';
+import { JSNS } from '@zextras/carbonio-ui-commons';
 import { AccountSettingsPrefs, soapFetchV2 } from '@zextras/carbonio-ui-soap-lib';
-
-import type { FolderSortOrder, SortDirection } from '../types';
-import { FILTER_OPTIONS, SORTING_OPTIONS } from '../constants';
 import { TFunction } from 'i18next';
 
-/**
- * Returns sortType, sortDirection and sortOrder for the given folder
- *
- * @param folderId
- * @param prefSortOrder
- *
- * returns an object containing
- * sortType: the sort type for the given folder,
- * sortDirection: the sort direction for the given folder,
- */
-
-const fallbackSortOrder: FolderSortOrder = {
-	sortType: 'date',
-	sortDirection: 'Desc' as SortDirection
-};
-
-const trashFolderSortOrder: FolderSortOrder = {
-	sortType: 'changeDate',
-	sortDirection: 'Desc' as SortDirection
-};
+import { FILTER_OPTIONS, SORTING_OPTIONS } from '../constants';
+import { findFolderEntry } from './parseMessageSortingOptions';
 
 export const getFilterQuery = (filter: string | undefined, folderId: string): string => {
 	if (!filter) return `inId:"${folderId}"`;
@@ -47,48 +26,6 @@ export const getFilterQuery = (filter: string | undefined, folderId: string): st
 			return `inId:"${folderId}"`;
 	}
 };
-
-function findFolderEntry(
-	prefSortOrder: string,
-	folderId: string
-): { currentFolder: string | undefined; parameters: string[] | undefined } {
-	if (!folderId || !prefSortOrder) return { currentFolder: undefined, parameters: undefined };
-
-	const folders = prefSortOrder.split(',');
-	const currentFolder = folders.find((folder) => folder.startsWith(`${folderId}:`));
-	if (!currentFolder) return { currentFolder: undefined, parameters: undefined };
-
-	const parameters = currentFolder.replace(',BDLV', '').replace(`${folderId}:`, '').split('-');
-
-	return { currentFolder, parameters };
-}
-
-export function parseMessageSortingOptions(
-	folderId: string,
-	prefSortOrder?: string
-): FolderSortOrder {
-	const isTrashFolder = isTrash(folderId);
-	const defaultSortOrder = isTrashFolder ? trashFolderSortOrder : fallbackSortOrder;
-
-	if (!prefSortOrder || !folderId) {
-		return defaultSortOrder;
-	}
-	const { parameters } = findFolderEntry(prefSortOrder ?? '', folderId);
-	if (parameters?.length === 2) {
-		return {
-			sortType: parameters[0],
-			sortDirection: parameters[1] as SortDirection
-		};
-	}
-	if (parameters?.length === 3) {
-		return {
-			sortType: parameters[0],
-			sortDirection: parameters[1] as SortDirection,
-			filterType: parameters[2]
-		};
-	}
-	return defaultSortOrder;
-}
 
 function modifySettingString(
 	zimbraPrefSortOrder: string,
