+++ src/views/app/folder-panel/conversations/conversation-list.tsx
@@ -12,10 +12,10 @@ import { map } from 'lodash';
 import { useParams } from 'react-router-dom';
 
 import { ConversationShortcutsRegister } from './conversation-shortcuts-register';
+import { parseMessageSortingOptions } from '../../../../helpers/parseMessageSortingOptions';
 import type { FolderPanelRouteParams } from '../../../../types/routes';
 import { API_REQUEST_STATUS, LIST_LIMIT } from 'constants/index';
 import { getFolderIdParts } from 'helpers/folders';
-import { parseMessageSortingOptions } from 'helpers/sorting';
 import { useConversationListByFolder } from 'hooks/use-conversations-list-by-folder';
 import { useMultipleSelection } from 'hooks/use-multiple-selection';
 import { ConversationListComponent } from 'views/app/folder-panel/conversations/conversation-list-component';
